Enemy Attack
Made enemy attack. Honestly, this attack is probably too accurate--it calculates a fire angle based on the current velocity of the player, meaning they player is guarenteed to get hit if they are moving in a straight line. It's cool, but possible unfair to the player. i COULD however, use this code to make an automated turret component, so if I do end up scrapping this, it was still good use of my time.