FUCKED IT UP (read desc)
So I actually made a lot of progress, I just named this Commit that because there is a glaring issue I created: changing the stats of a ScriptableObject changes the stats for EVERY component with that scriptable object. So I need to allow the ShipComponent class instance to store post-buff data, NOT the ScriptableObject.