STILL FUCKED UP LOL!!!
Ok so I learned that trying to access an instance of a component through the Grid in ShipBuildData is BAD because that Grid doesn't actually store the information of the instances, it stores the information of the PREFABS!!! So when applying buffs, it actually just applies to the prefabs... So I'm gonna have to make it possible to access an instance of a component through ShipBuildData somehow, but I've been working on this for so long that I can't bring myself to do it right now. Good luck future self.